fix: fix player argument parser not getting all players
