+++ versions/1.8.9-forge/api/1.8.9-forge.api
@@ -105,12 +105,20 @@ public class cc/polyfrost/oneconfig/utils/commands/PlatformCommandManagerImpl : 
 	public fun createCommand (Lcc/polyfrost/oneconfig/utils/commands/CommandManager$OCCommand;)V
 }
 
-public class cc/polyfrost/oneconfig/utils/commands/arguments/PlayerArgumentParser : cc/polyfrost/oneconfig/utils/commands/arguments/ArgumentParser {
+public class cc/polyfrost/oneconfig/utils/commands/arguments/OldPlayerArgumentParser : cc/polyfrost/oneconfig/utils/commands/arguments/ArgumentParser {
 	public fun <init> ()V
 	public fun complete (Ljava/lang/String;Ljava/lang/reflect/Parameter;)Ljava/util/List;
 	public static fun getMatchingPlayers (Ljava/lang/String;)Ljava/util/List;
 	public synthetic fun parse (Ljava/lang/String;)Ljava/lang/Object;
 	public fun parse (Ljava/lang/String;)Lnet/minecraft/entity/player/EntityPlayer;
+}
+
+public class cc/polyfrost/oneconfig/utils/commands/arguments/PlayerArgumentParser : cc/polyfrost/oneconfig/utils/commands/arguments/ArgumentParser {
+	public fun <init> ()V
+	public fun complete (Ljava/lang/String;Ljava/lang/reflect/Parameter;)Ljava/util/List;
+	public static fun getMatchingPlayers (Ljava/lang/String;)Ljava/util/List;
+	public fun parse (Ljava/lang/String;)Lcom/mojang/authlib/GameProfile;
+	public synthetic fun parse (Ljava/lang/String;)Ljava/lang/Object;
 }
 
 public final class cc/polyfrost/oneconfig/utils/dsl/DSLsKt {
