+++ versions/src/main/java/cc/polyfrost/oneconfig/utils/commands/PlatformCommandManagerImpl.java
@@ -31,6 +31,7 @@ import cc.polyfrost.oneconfig.libs.universal.UChat;
 import cc.polyfrost.oneconfig.utils.StringUtils;
 import cc.polyfrost.oneconfig.utils.commands.annotations.Description;
 import cc.polyfrost.oneconfig.utils.commands.annotations.Greedy;
+import cc.polyfrost.oneconfig.utils.commands.arguments.OldPlayerArgumentParser;
 import cc.polyfrost.oneconfig.utils.commands.arguments.PlayerArgumentParser;
 import net.minecraft.command.CommandBase;
 import net.minecraft.util.BlockPos;
@@ -50,6 +51,7 @@ import net.minecraftforge.client.ClientCommandHandler;
 public class PlatformCommandManagerImpl extends PlatformCommandManager {
 
     static {
+        INSTANCE.addParser(new OldPlayerArgumentParser());
         INSTANCE.addParser(new PlayerArgumentParser());
     }
