+++ versions/src/main/java/cc/polyfrost/oneconfig/utils/commands/arguments/PlayerArgumentParser.java
@@ -1,9 +1,10 @@
-//#if MC<=11202
 package cc.polyfrost.oneconfig.utils.commands.arguments;
 
 import com.google.common.collect.Lists;
+import com.mojang.authlib.GameProfile;
 import net.minecraft.client.Minecraft;
-import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.client.network.NetworkPlayerInfo;
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -11,23 +12,24 @@ import java.lang.reflect.Parameter;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class PlayerArgumentParser extends ArgumentParser<EntityPlayer> {
+/**
+ * The player argument parser. Returns a {@link GameProfile}.
+ */
+public class PlayerArgumentParser extends ArgumentParser<GameProfile> {
     @Nullable
     @Override
-    public EntityPlayer parse(@NotNull String arg) throws Exception {
+    public GameProfile parse(@NotNull String arg) throws Exception {
         return getMatchingPlayers(arg).stream().findFirst().orElseThrow(() -> new Exception("No player found"));
     }
 
-    public static List<EntityPlayer> getMatchingPlayers(String arg) {
+    public static List<GameProfile> getMatchingPlayers(String arg) {
         if (Minecraft.getMinecraft().theWorld == null) return Lists.newArrayList();
-        return Minecraft.getMinecraft().theWorld.getPlayers(EntityPlayer.class,
-                player -> player.getName().equals(arg) || player.getUniqueID().toString().equals(arg));
+        return Minecraft.getMinecraft().getNetHandler().getPlayerInfoMap().stream().map(NetworkPlayerInfo::getGameProfile).filter(gameProfile -> StringUtils.startsWithAny(gameProfile.getName().toLowerCase(), "!", arg.toLowerCase())).collect(Collectors.toList());
     }
 
     @NotNull
     @Override
     public List<String> complete(String current, Parameter parameter) {
-        return getMatchingPlayers(current).stream().map(EntityPlayer::getName).collect(Collectors.toList());
+        return getMatchingPlayers(current).stream().map(GameProfile::getName).collect(Collectors.toList());
     }
 }
-//#endif
